KDD-072 P-b2: predecessor bound into the Phase-4 sign-hash (six callers, one implementation)
- One GetSignHash(predecessor) implementation, predecessor appended iff
  non-zero (v1 preimage 144 bytes byte-identical, v2 176). The message wire
  form is UNCHANGED; the predecessor rides the preimage, not the
  serialization.
- Six callers, not three. Recon found 3 by reading (Build, Receive check-7,
  V7g); the build found 3 more - transport R4 (ptx_dkg_net.cpp) and the
  debug builder x2 (rpc/ptx.cpp) - surfaced by the compiler because
  GetSignHash has no default parameter. A defaulted signature would have
  shipped callers computing mismatched preimages as SILENT wrong-preimage
  bugs; no-default made them compile errors. Second time this arc the
  compiler backstopped the read (P-a's nVersion was the first).
- Section-3 hole test isolates the sign-hash as the cause: control passes
  (genuine v1 verifies), attack fails with exactly "ptxdkg-bad-premit-sig"
  (V7a-V7f pass by construction, V7g the isolated cause), inversion strips
  only the binding and the attack SUCCEEDS with "the KDD-072 section-3
  unsigned-predecessor hole is OPEN". The specific reject string and
  control-passes-first are the isolation.
- Four inversions RED at four distinct caller sites (INV-1 validator, INV-2
  sign, INV-3 receive, INV-4a Phase-5 derivation) - the shared-preimage/
  divergent-view design proven from every caller.
- Fresh-path golden 761090b8... captured by execution at pre-edit HEAD,
  reproduced byte-for-byte; INV-4b (always-append) caught by it, proving it
  is the guard against uniform preimage drift.
- V7g validator-side branch lands in P-b2 deliberately - dormant-safe: no v2
  reaches VerifyPremits via CheckPTXDKGTx while the gate rejects v2 (P-b3),
  unit tests call the pure function directly. Required for the hole test.
  Not scope creep.
- Phase-5 dormant consumers: BuildPTXDKGTx derives nVersion from
  session.predecessor, ClosePhase5 stores role=PENDING iff rotation
  (KDD-070 section-8 feed). Both dormant - the session field is never fed
  in P-b2.
- Debug builder passes pl.predecessor_quorum_hash - KDD-073 reconstruction
  site #3, already lockstep-correct before P-b3 needs it.
- Triple-locked dormancy: gate rejects v2 (CI-pinned), setter
  present-but-unfed (only SetNull at ptx_formation.cpp:360, grep-proven),
  fresh preimage golden. Plus: a hypothetical rotation FINALIZE has its
  commitment refused by the KDD-058-A store (AddAndRelay runs the version
  gate).
- 323 -> 329 green. 100% pure-function, no bf arm (real rotation premits
  need the P-b3 gate + P-b6 trigger).
- Arc: P-b2 of P-b1 -> P-b2 -> P-b4 -> P-b3 -> P-b5 -> P-b6. P-b4 (ODC-042
  as-of fix) next, before P-b3.

+++ src/ptx/test/ptx_dkg_serialization_tests.cpp
@@ -288,15 +288,17 @@ BOOST_AUTO_TEST_CASE(s4_phase4_roundtrip)
     memset(msg.group_pk_bytes, 0xAB, 48);
     buf.assign(32, 0x66);
     msg.vvec_hash = uint256(buf);
-    msg.sig = sk.Sign(msg.GetSignHash());
+    // KDD-072 P-b2: fresh (zero-predecessor) view — this row is wire-form
+    // stability; the predecessor rides the sign-hash, never the serialization.
+    msg.sig = sk.Sign(msg.GetSignHash(uint256()));
 
     auto bytes = SerBytes(msg);
     PTXDKGPhase4Msg out = DeserBytes<PTXDKGPhase4Msg>(bytes);
 
     BOOST_CHECK(SerBytes(out) == bytes);
-    BOOST_CHECK(out.GetSignHash() == msg.GetSignHash());
+    BOOST_CHECK(out.GetSignHash(uint256()) == msg.GetSignHash(uint256()));
     BOOST_CHECK(memcmp(out.group_pk_bytes, msg.group_pk_bytes, 48) == 0);
-    BOOST_CHECK(out.sig.VerifyInsecure(sk.GetPublicKey(), out.GetSignHash()));
+    BOOST_CHECK(out.sig.VerifyInsecure(sk.GetPublicKey(), out.GetSignHash(uint256())));
 }
 
 // ---------------------------------------------------------------------------
